requirements.txt
# Core Data Science Libraries
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
shap>=0.45.0

# Utilities
joblib>=1.3.0

# Kaggle Dataset Access
kagglehub>=0.2.0

# Jupyter (optional, for running notebooks)
jupyter>=1.0.0
jupyterlab>=4.0.0

# Configuration
pyyaml>=6.0

# Memory Profiling
memory-profiler>=0.61.0

# Testing
pytest>=7.0.0

# Additional sklearn utilities (implicit but explicit for clarity)
scipy>=1.11.0 # Required by scikit-learn
